refactor: register durable review continuation (PAN-3135)
Keep review-status cycle-safe by registering the dashboard-owned verification,

push, and review continuation through a neutral host callback.
