+++ src/dashboard/server/routes/workspaces/review-pipeline.ts
@@ -30,8 +30,10 @@ import { getReleaseSetSync } from '../../../../lib/release-set.js';
 import { getCachedConflictGateMergeability } from '../../../../lib/cloister/conflict-gate.js';
 import { transitionIssueToInReview, spawnRun } from '../../../../lib/agents.js';
 import { runVerificationForIssue } from '../../../../lib/cloister/verification-runner.js';
+import { registerDurableReviewPipelineHandler } from '../../../../lib/cloister/durable-review-pipeline.js';
 import { requestReviewPipeline } from '../../../../lib/cloister/request-review-pipeline.js';
 import { jsonResponse } from '../../http-helpers.js';
+import { startDashboardDurableReviewPipeline } from '../../services/durable-review-pipeline.js';
 import { httpHandler } from '../http-handler.js';
 import {
   getProjectPath,
@@ -47,7 +49,7 @@ import {
 
 const execAsync = promisify(exec);
 const MAX_AUTO_REQUEUE = 25;
-
+registerDurableReviewPipelineHandler(startDashboardDurableReviewPipeline);
 /** Safe `.message` read for caught values of unknown shape. */
 const errorMessage = (e: unknown): string | undefined => e instanceof Error ? e.message : undefined;
 
@@ -119,7 +121,6 @@ async function getDirtyWorkspaceErrorForReviewRequest(
   }
 }
 // ─── Route: POST /api/review/:issueId/trigger ─────────────────────────────
-
 const postWorkspaceReviewRoute = HttpRouter.add(
   'POST',
   '/api/review/:issueId/trigger',
@@ -406,7 +407,6 @@ const postWorkspaceReviewRoute = HttpRouter.add(
   }))
 );
 // ─── Route: POST /api/review/:issueId/request ─────────────────────
-
 const postWorkspaceRequestReviewRoute = HttpRouter.add(
   'POST',
   '/api/review/:issueId/request',
